deleted some useless logs, added minute for start and end timeof posting, fixed interval of posting bug

+++ utils.py
@@ -7,7 +7,7 @@ import logging
 import subprocess
 import asyncio
 import random
-from datetime import datetime, timedelta
+from datetime import datetime, time, timedelta
 
 import aiohttp
 
@@ -18,6 +18,8 @@ from config import (
     CLIP_NAME,
     CLIP_DESCRIPTION,
     CLEANUP_INTERVAL_HOURS,
+    POSTING_INTERVAL,
+    LAST_TIME_POST,
     USE_YTDLP,
     YTDLP_PATH
 )
@@ -75,22 +77,36 @@ async def post_clip_to_wall():
     """
     Публицакия клипов на стене группы
     """
-    clip_path = get_random_clip()
-    print(f"Выбран клип для загрузки: {clip_path}")
 
-    try:
-        video_uploader = VideoUploader(api)
-        attachment = await video_uploader.upload(
-            file_source=clip_path,
-            name=CLIP_NAME,
-            description=CLIP_DESCRIPTION,
-            group_id=GROUP_ID,
-            wallpost=True
-        )
-        os.remove(clip_path)
-        logger.info(f"Клип успешно опубликован на стене сообщества (ID группы: {GROUP_ID}).")
-    except Exception as e:
-        logger.error("Ошибка при загрузке клипа на стену: %s", e)
+    if (datetime.now() - LAST_TIME_POST >= timedelta(hours = POSTING_INTERVAL)):
+        clip_path = get_random_clip()
+        print(f"Выбран клип для загрузки: {clip_path}")
+
+        try:
+            video_uploader = VideoUploader(api)
+            attachment = await video_uploader.upload(
+                file_source=clip_path,
+                name=CLIP_NAME,
+                description=CLIP_DESCRIPTION,
+                group_id=GROUP_ID,
+                wallpost=True
+            )
+            os.remove(clip_path)
+            with open('config.py', 'r+') as f:
+                lines = f.readlines()
+                f.seek(0)
+                for line in lines:
+                    if line.startswith('LAST_TIME_POST'):
+                        f.write(f"LAST_TIME_POST = datetime({datetime.now().year}, {datetime.now().month}, {datetime.now().day}, {datetime.now().hour}, {datetime.now().minute}, {datetime.now().second})\n")
+                    else:
+                        f.write(line)
+                f.truncate()
+            logger.info(f"Клип успешно опубликован на стене сообщества (ID группы: {GROUP_ID}).")
+        except Exception as e:
+            logger.error("Ошибка при загрузке клипа на стену: %s", e)
+    else: 
+        logger.error("Слишком рано для поста.")
+        return
 
 
 async def download_with_ytdlp(video_url, output_path):
